Organize the tests so they’re 1-1 between files and tests.
Also allow specs to be loaded with `rake test`.
